+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+* Add additional Asset Manager test helper methods
+
 # 57.0.0
 
 * Pass `LINK_CHECKER_API_BEARER_TOKEN` in Link Checker API requests if present.